Separate IP address observer into separate script
Before, ip_addresss_observer.py simultaneously served two different functions, one to provide imports used in charm.py and one to call charm.py itself (via juju-run).

Now, those two functions are separated.

Fixes issue where ops import causes `juju-run` script to fail because ops is not available from /usr/bin/python3 when using a full virtual environment (needed for charmcraft 3 poetry plugin migration)—now `juju-run` script does not import ops since it never needed it

Context: https://chat.canonical.com/canonical/pl/drhbrt84y3yc7xwc5431a1o5de
